% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsim.R
\docType{package}
\name{microsim}
\alias{microsim}
\title{microsim package}
\description{
Simulator Package with a collection of basic community models that outputs count data of microbial communities.
}
\author{
Emma Gheysen
}
