% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutate_microsats.R
\name{mutate_microsats}
\alias{mutate_microsats}
\title{A function to mutate microsatellites}
\usage{
mutate_microsats(n_mutations, mutation_model = "smm")
}
\arguments{
\item{n_mutations}{A vector indicating the total number of mutations along
each branch of a tree}

\item{mutation_model}{A character string indicating the mutation model to use.
Currently, only the stict stepwise mutation model is implemented ('smm').
Default is 'smm'}
}
\description{
A function to mutate microsatellites
}

