% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomA.R
\name{randomA}
\alias{randomA}
\title{Generate random uniform interaction matrix}
\usage{
randomA(N = 100, d = -0.5, min.strength = -0.5, max.strength = 0.5, c = 0.02)
}
\arguments{
\item{N}{number of species}

\item{d}{diagonal values (should be negative)}

\item{min.strength}{minimal off-diagonal interaction strength}

\item{max.strength}{maximal off-diagonal interaction strength}

\item{c}{connectance (interaction probability)}
}
\description{
Generate random simplified interaction matrix from a uniform distribution.
}
\note{
Simplified version of generateA() function from seqtime package
}
\examples{
high_inter_A <- randomA(10, d = -0.4, min.strength = -0.8, max.strength = 0.8, c = 0.5)
low_inter_A <- randomA(10, c = 0.01)
}
\seealso{
seqtime generateA()
}
