% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rgeom_dist}
\alias{rgeom_dist}
\title{Generate random numbers from the DiRienzo Symmetric Geometric distribution}
\usage{
rgeom_dist(n, p = NULL, sigma2 = NULL)
}
\description{
Generate random numbers from the DiRienzo Symmetric Geometric distribution
}

