% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_sample.R
\name{rmse_sample}
\alias{rmse_sample}
\title{Lowest-RMSE sample from time series}
\usage{
rmse_sample(spab)
}
\arguments{
\item{spab}{Species abundances matrix with OTUs in the rows and the timepoints as columns.}
}
\value{
A sample vector with the length equal to the number of rows of given input species abundances matrix.
}
\description{
Computes the Root Mean Square Error between subsequent generations (columns)
in given species abundances matrix (time series) as a measure of convergence.
The timepoint with the lowest RMSE-difference with respect to the previous timepoint
is then extracted and a normalised sample is returned (proportions).
}
