% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_migrn.R
\name{write_migrn}
\alias{write_migrn}
\title{Create input file for Migrate-N}
\usage{
write_migrn(genotypes, filename, n_ind, n_pop, n_loc, title)
}
\arguments{
\item{genotypes}{A matrix outputted by sim_microsats function.}

\item{filename}{A character string indicating the filename to be used for saving
the input file.}

\item{n_ind}{An integer indicating the number of sampled diploid individuals
if n_pop = 1, or a vector of integers indicating the number of
diploid individuals sampled in each population in n_pop > 1.}

\item{n_pop}{An integer indicating the number of populations simulated.}

\item{n_loc}{An integer indicating the number of loci simulated.}

\item{title}{A character string to be used as the title of the input file,
only used if the format allows it.}
}
\description{
Not a function for general use, please use save_microsimr()
             instead
}

