% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPhyloseq.R
\name{asPhyloseq}
\alias{asPhyloseq}
\title{Microbiome Phyloseq wrapper for community samples}
\usage{
asPhyloseq(S, ...)
}
\arguments{
\item{S}{Each parameter given to the function is a sample vector or matrix to be merged
into one phyloseq object of samples.
The vectors/matrices must all contain the same number of species in order to be merged.
A phyloseq object requires at least a matrix of min. 2 columns,
therefore an error is thrown if only a vector is given without additional vector/matrix objects.}

\item{...}{vector(s)/matrice(s): see description of S}
}
\value{
A microbiome phyloseq class object with species/OTUs as rows and samples as columns.
}
\description{
Takes in samples in vector or matrix format
and merges them to a microbiome phyloseq class object.
}
\details{
Requires phyloseq package (Bioconductor). Function will install package when not installed yet.
}
\examples{
\dontrun{
A = powerlawA(n = 80, alpha = 1.6),
glv_timeseries <- glv(N = 80, A = A, tend = 100)
soi_sample <- rmse_sample(soi(N = 80, I = 1000, A = A, tend = 1000))
physeq <- asPhyloseq(glv_timeseries, soi_sample)
}
}
