% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glv.R
\name{glv}
\alias{glv}
\title{Simulate time series with the generalized Lotka-Volterra model}
\usage{
glv(
  N = 4,
  A,
  growth_rates = runif(N),
  com = runif(N),
  tend = 100
)
}
\arguments{
\item{N}{species number}

\item{A}{interaction matrix}

\item{growth_rates}{growth rates}

\item{com}{initial abundances}

\item{tend}{final time point}
}
\value{
a matrix with species abundances as rows and time points as columns, column names give time points
}
\description{
Simulate a community time series using the generalized Lotka-Volterra model, defined as
\eqn{\frac{dx}{dt}=x(b+Ax)}, where x is the vector of species abundances, A is the interaction matrix
and b the vector of growth rates.
}
\examples{
glv(N=4,generateA(4))
}
