% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubbell.R
\name{hubbell}
\alias{hubbell}
\title{Hubbell community simulation}
\usage{
hubbell(
  N = NULL,
  M = NULL,
  I,
  d,
  m,
  com = NULL,
  pbirth = NULL,
  pmigr = NULL,
  tskip = 0,
  tend,
  norm = FALSE
)
}
\arguments{
\item{N}{Amount of different species initially in the local community}

\item{M}{Amount of different species in the metacommunity, including those of the local community}

\item{I}{Fixed amount of individuals in the local community}

\item{d}{Fixed amount of deaths of local community individuals in each generation}

\item{m}{Immigration rate: the probability that a death in the local community is
replaced by a migrant of the metacommunity rather than by the birth of a local community member}

\item{com}{initial community vector of length M where the ith element is the amount of individuals of the ith species}

\item{pbirth}{probabilities of birth}

\item{pmigr}{probabilities of migration}

\item{tskip}{Nr of generations that should not be included in the outputted species
abundance matrix.}

\item{tend}{Nr of simulations to be simulated.}

\item{norm}{logical to indicate whether the time series should be returned
with the abundances as proportions (norm = TRUE) or the raw counts
(norm = FALSE, default)}
}
\value{
matrix with species abundances as rows and time points as columns
}
\description{
Neutral species abundances simulation according to the Hubbell model.
}
\details{
For more theoretical background see: Rosindell, James et al. “The unified neutral theory of biodiversity and biogeography at age ten.” Trends in ecology & evolution vol. 26,7 (2011).
}
\examples{
hubbell(N = 8, M = 10, I = 1000, d = 50, m = 0.02, tend = 100)
}
