% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_microsats.R
\name{mutate_microsats}
\alias{mutate_microsats}
\title{A function to mutate microsatellites}
\usage{
mutate_microsats(n_mutations, mutation_model = "smm", p_single = 0.8,
  sigma2 = 50)
}
\arguments{
\item{n_mutations}{A vector indicating the total number of mutations along
each branch of a tree}

\item{mutation_model}{A character string indicating the mutation model to use.
Currently, only the strict stepwise mutation model of Ohta and Kimura (1973) ('smm'),
and the DiRienzo et al. (1994) two-phase model ('tpm') are implemented.
Default is 'smm'}

\item{p_single}{Probability of a single-step mutation to be used in the 'tmp' model}

\item{sigma2}{Variance in allele size to be used in the 'tpm' model}
}
\description{
A function to mutate microsatellites
}
\details{
A tree is first simulated using 'ms'. Mutations are simulated along the
branches of the tree following a Poisson distribution with lambda proportional
to branch length times theta (4Nmu). This first part is done in \link{sim_microsats}.

Here, the number of mutations at each branch is transformed to either a loss
or gain in number of repeats. In the 'smm' model, each mutation represents either
a loss or gain of a single repeat unit with equal probability. In the 'tpm'
model, with probability \code{p_single} a mutation represents either a
gain or loss of a single repeat, and with probability (1 - \code{p_single})
the gain/loss is larger following a symmetric geometric distribution.

Please refer to the vignette to see a deeper explanation, and test of each
model.
}
\references{
Di Rienzo, A., Peterson, A. C., Garza, J. C., Valdes, A. M., Slatkin, M., & Freimer, N. B. (1994). Mutational processes of simple-sequence repeat loci in human populations. Proceedings of the National Academy of Sciences of the United States of America, 91(8), 3166–3170.

Ohta, T., & Kimura, M. (2007). A model of mutation appropriate to estimate the number of electrophoretically detectable alleles in a finite population. Genetical Research, 89(5-6), 367–370. http://doi.org/10.1017/S0016672308009531
}

