% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_microsats.R
\name{sim_microsats}
\alias{sim_microsats}
\title{Simulate micrasatellite data using the coalescent}
\usage{
sim_microsats(theta, n_ind, n_loc, n_pop = NULL, mutation_model = "smm",
  ancestral_allele_size = 80, ms_options = NULL)
}
\arguments{
\item{theta}{The per locus mutation rate.}

\item{n_ind}{An integer indicating the number of diploid individuals to be
simulated, or a vector indicating how many individuals to sample
per simulated population}

\item{n_loc}{An integer indicating the number of loci to be simulated.}

\item{n_pop}{An integer indicating the number of populations to be simulated.
If NULL, then assume n_pop = 1.}

\item{mutation_model}{A character string indicating which mutation model to use.
Currently, only the strict stepwise mutation model ('smm') is implemented.
Default is 'smm' (see Details).}

\item{ancestral_allele_size}{An integer indicating the ancestral allele
size (i.e, the count of number of repeats of the ancestral allele).
Default is 80 (see Details).}

\item{ms_options}{A string with additional options to pass on to ms.}
}
\description{
Simulate micrasatellite data using the coalescent
}
\examples{
# generate a simulated dataset for a single population with theta = 5,
# from which 100 loci were genotyped for 20 individuals
simd_data_1pop <- microsimr::sim_microsats(theta = 5,
                                           n_ind = 20,
                                           n_loc = 100,
                                           n_pop = 1)

# generate a simulated dataset for a three populations with theta = 12,
# from which 100 loci were genotyped for 20 individuals each sampled from
# three populations with 20 migrants per generation

simd_data_1pop <- microsimr::sim_microsats(theta = 12,
                                           n_ind = c(20, 20, 20),
                                           n_loc = 100,
                                           n_pop = 3,
                                           ms_options = "-I 3 40 40 40 20")
# Notice the 'ms_options' parameter, the -I indicates to model an island
# population structure model. The three 40's indicate to sample 40 chromosomes
# from each of the three modeled populations, and the 20 tells ms to simulate
# populations exchanging 20 migrants per generation. Additional options,
# including population growth, assymetric migration, isolation, etc. can
# be modeled by using the right combination of parameters. These are laid
# out in ms's manual. Examples can be added depending on user's needs. Please
# request additional examples through the GitHub issues page.
}

