% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_microsats.R
\name{sim_microsats}
\alias{sim_microsats}
\title{Simulate micrasatellite data using the coalescent}
\usage{
sim_microsats(theta, n_ind, n_loc, ancestral_allele_size = 80,
  ms_options = NULL)
}
\arguments{
\item{theta}{The per locus mutation rate.}

\item{n_ind}{An integer indicating the number of diploid individuals to be
simulated.}

\item{n_loc}{An integer indicating the number of loci to be simulated.}

\item{ancestral_allele_size}{An integer indicating the ancestral allele
size. Default is 80 (see Details).}

\item{ms_options}{A string with additional options to pass on to ms.}
}
\description{
Simulate micrasatellite data using the coalescent
}

